% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_suspected_data.R
\name{add_suspected_rows}
\alias{add_suspected_rows}
\title{Identify rows of a data set to be marked as suspicious}
\usage{
add_suspected_rows(x, reason, suspect_data = NULL)
}
\arguments{
\item{x}{a subset of rows from the tibble of interest}

\item{reason}{a character string describing why this particular subset is
suspect}

\item{suspect_data}{a specifically structure tibble

If suspect_data is NULL, a new suspect_data tibble will be constructed and
the identified data added to that structure.}
}
\value{
a tibble with the suspect data structure
}
\description{
Given a common tibble of interest, a subset of the data can be identified as
suspicious and added to the suspect_data structure for analysis and
visualization.
}
\examples{
df <- create_suspicious_data()
suspect_data <-
  df \%>\%
  filter(cut == "Fair") \%>\%
  add_suspected_rows(reason = "Cut was not good or better")

}
