% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjudicate_suspect_data.R
\name{adjudicate_suspect_data}
\alias{adjudicate_suspect_data}
\title{Adjudicate suspected data removing it from the original data set}
\usage{
adjudicate_suspect_data(ds, suspect_data)
}
\arguments{
\item{ds}{tibble containing the original data set}

\item{suspect_data}{a suspect_date object to be applied to the original data
set}
}
\value{
an adjudicated_data object containing the original and adjudicated
data sets and associated metadata.
}
\description{
Per the contents of the suspect_data, remove the rows and columns identified.
}
\examples{
original_data <- create_suspicious_data()
suspect_data <-
  original_data \%>\%
  filter(cut == "Fair") \%>\%
  add_suspected_rows(reason = "Cut was not good or better")

suspect_data <-
  original_data \%>\%
  filter(price < 5000) \%>\%
  add_suspected_rows(reason = "Price was less than 5000", suspect_data)

suspect_data <-
  original_data \%>\% filter(table < 57) \%>\%
  add_suspected_rows(reason = "Table was less than 57", suspect_data)

suspect_data <-
  original_data \%>\%
  select(clarity, y) \%>\%
  add_suspected_columns(reason = "These columns are hinky", suspect_data)

suspect_data <-
  original_data \%>\%
  select(depth) \%>\%
  add_suspected_columns(reason = "This column is not deep enough", suspect_data)

adjudicated_data <- adjudicate_suspect_data(original_data, suspect_data)
}
