% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missingness_by_var.R
\name{as_lith_tbl}
\alias{as_lith_tbl}
\title{Create a copy of a tibble replacing all of the elements with a binary missing
or present factor.}
\usage{
as_lith_tbl(ds, reference_column = NULL)
}
\arguments{
\item{ds}{a tibble to be developed as a data lithograph}

\item{reference_column}{an unquoted character string designating a reference
column to be preserved for sorting and future visualization. If no column
is designated, the function will use the row_identifier column.}
}
\value{
a tibble comprised of the original columns containing present/missing
values and the reference column in its original state.
}
\description{
The function takes a tibble composed of arbitrary data types and creates a
copy of the structure but replaces all of the values with either 'present' or
'missing'. Something like a photographic lithograph.
}
\details{
In addition to replace the values, two columns are added: row_identifier and
reference_column. These are used to allow for the transformed data to be
incorporated into subsequent analysis and visualization. If not reference
column is provided, the row_identifier column will be used as the reference_column.
}
\examples{
ds <- create_suspicious_data()
as_lith_tbl(ds, reference_column = carat) \%>\% head()
as_lith_tbl(ds, reference_column = key_date) \%>\% head()
as_lith_tbl(ds) \%>\% head()
}
