% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inquisitive-package.R
\docType{package}
\name{inquisitive-package}
\alias{inquisitive}
\alias{inquisitive-package}
\title{inquisitive: A collection of exploratory data analysis functions used to identify, quantify, and visualize data quality issues}
\description{
This package curates a set of data analysis functions focused on profiling data for machine learning and providing feedback to upstream business process and data owners to facilitate the improvement of data quality and completeness.
}
\author{
\strong{Maintainer}: John Clouse \email{john@clouse.io} (\href{https://orcid.org/0000-0002-0369-2116}{ORCID})

}
\keyword{internal}
