% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missingness_by_var.R
\name{plot_missingness_by_var}
\alias{plot_missingness_by_var}
\title{Plot missingness by variable}
\usage{
plot_missingness_by_var(
  ds,
  reference_column = NULL,
  number_of_axis_lables = 25,
  date_format_fun = function(x) {
     format(x, "\%Y-\%m")
 }
)
}
\arguments{
\item{ds}{a data frame to be plotted}

\item{reference_column}{an unquoted column name}

\item{number_of_axis_lables}{an integer specifying maximum number of labels
for the x axis}

\item{date_format_fun}{a function specifying the formatting of of the x axis
when the column selected is of type Date.}
}
\description{
Build a ggplot depicting missingness in a data frame. The data is first
turned into a lithographic copy of the frame with only 'missing' or 'present'
by \code{\link{as_lith_tbl}}.
}
\details{
If a reference column is specified, the data frame will be sorted by the
reference column prior to plotting
}
\examples{
ds <- inquisitive::create_suspicious_data()
plot_missingness_by_var(ds, reference_column = key_date)
plot_missingness_by_var(ds, reference_column = key_date, date_format_fun =
 function(x) {format(x, "\%b- \%Y")})
plot_missingness_by_var(ds, reference_column = carat)
plot_missingness_by_var(ds)
}
