% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_suspect_data.R
\name{plot_suspect_data}
\alias{plot_suspect_data}
\title{Plot the effects of removing suspected data against the original data set}
\usage{
plot_suspect_data(suspect_ds, adjudicated_ds)
}
\arguments{
\item{suspect_ds}{a suspect_data object}

\item{adjudicated_ds}{an adjudicated data object}
}
\value{
ggplot object
}
\description{
Plot the effects of removing suspected data against the original data set
}
\examples{
df <- create_suspicious_data()
suspect_data <-
  df \%>\%
  filter(cut == "Fair") \%>\%
  add_suspected_rows(reason = "Cut was not good or better")

suspect_data <-
  df \%>\%
  filter(price < 5000) \%>\%
  add_suspected_rows(reason = "Price was less than 5000", suspect_data)

suspect_data <-
  df \%>\%
  filter(table < 57) \%>\%
  add_suspected_rows(reason = "Table was less than 57", suspect_data)

suspect_data <-
  df \%>\%
  select(clarity, y) \%>\%
  add_suspected_columns(reason = "These columns are hinky",
                        suspect_data)

adjudicated_data <- adjudicate_suspect_data(df, suspect_data)

plot_suspect_data(suspect_ds = suspect_data, adjudicated_ds = adjudicated_data)
}
