% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_suspected_data.R
\name{suspect_data_constructor}
\alias{suspect_data_constructor}
\title{Constructor function for suspected data tibble structure}
\usage{
suspect_data_constructor()
}
\value{
A tibble structured to support suspecting data
}
\description{
Suspected data identified by row or column is stored in a common structure
supporting independent collection of suspicious data prior to final adjudication.
}
\examples{
suspect_data_constructor()
}
