% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_suspicious_data.R
\name{create_suspicious_data}
\alias{create_suspicious_data}
\title{Create sample data for the inquisitive}
\usage{
create_suspicious_data()
}
\value{
A tibble.
}
\description{
A function to create a set of data useful for testing and evauating inquisitive tools.
}
